requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.9
openai==1.3.0
pydub==0.25.1
python-dotenv==1.0.0